assistant2: Restructure tools in preparation for adding more
